chore: refactor config
