+++ packages/core/src/benchmark/benches/recommended-oas3.bench.ts
@@ -1,10 +1,9 @@
 import { readFileSync } from 'fs';
 import { join as pathJoin, resolve as pathResolve } from 'path';
 import { lintDocument } from '../../lint';
-import { LintConfig } from '../../config/config';
+import { LintConfig, defaultPlugin } from '../../config';
 import { BaseResolver } from '../../resolve';
 import { parseYamlToDocument } from '../utils';
-import { defaultPlugin } from '../../config/builtIn';
 
 export const name = 'Validate with recommended rules';
 export const count = 10;
