+++ packages/core/src/config/config.ts
@@ -1,126 +1,37 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import { dirname } from 'path';
-import { red, blue, yellow, green } from 'colorette';
 import { parseYaml, stringifyYaml } from '../js-yaml';
-import { notUndefined, slash } from '../utils';
+import { slash } from '../utils';
+import { NormalizedProblem } from '../walk';
+import recommended from './recommended';
 import {
   OasVersion,
-  Oas3PreprocessorsSet,
   OasMajorVersion,
-  Oas3DecoratorsSet,
   Oas2RuleSet,
-  Oas2PreprocessorsSet,
-  Oas2DecoratorsSet,
   Oas3RuleSet,
 } from '../oas-types';
-import { ProblemSeverity, NormalizedProblem } from '../walk';
-import recommended from './recommended';
-import { NodeType } from '../types';
+
+import type { NodeType } from '../types';
+import type {
+  Api,
+  DecoratorConfig,
+  LintRawConfig,
+  Plugin,
+  PreprocessorConfig,
+  RawConfig,
+  Region,
+  ResolveConfig,
+  RuleConfig,
+} from './types';
+import { mergeExtends, resolvePlugins, resolvePresets } from './utils';
 
 export const IGNORE_FILE = '.redocly.lint-ignore.yaml';
 const IGNORE_BANNER =
   `# This file instructs Redocly's linter to ignore the rules contained for specific parts of your API.\n` +
   `# See https://redoc.ly/docs/cli/ for more information.\n`;
 
-export type RuleConfig =
-  | ProblemSeverity
-  | 'off'
-  | ({
-      severity?: ProblemSeverity;
-    } & Record<string, any>);
-
-export type PreprocessorConfig =
-  | ProblemSeverity
-  | 'off'
-  | 'on'
-  | ({
-      severity?: ProblemSeverity;
-    } & Record<string, any>);
-
-export type DecoratorConfig = PreprocessorConfig;
-
-export type LintRawConfig = {
-  plugins?: (string | Plugin)[];
-  extends?: string[];
-  doNotResolveExamples?: boolean;
-
-  rules?: Record<string, RuleConfig>;
-  oas2Rules?: Record<string, RuleConfig>;
-  oas3_0Rules?: Record<string, RuleConfig>;
-  oas3_1Rules?: Record<string, RuleConfig>;
-
-  preprocessors?: Record<string, PreprocessorConfig>;
-  oas2Preprocessors?: Record<string, PreprocessorConfig>;
-  oas3_0Preprocessors?: Record<string, PreprocessorConfig>;
-  oas3_1Preprocessors?: Record<string, PreprocessorConfig>;
-
-  decorators?: Record<string, DecoratorConfig>;
-  oas2Decorators?: Record<string, DecoratorConfig>;
-  oas3_0Decorators?: Record<string, DecoratorConfig>;
-  oas3_1Decorators?: Record<string, DecoratorConfig>;
-};
-
-export type PreprocessorsConfig = {
-  oas3?: Oas3PreprocessorsSet;
-  oas2?: Oas2PreprocessorsSet;
-};
-
-export type DecoratorsConfig = {
-  oas3?: Oas3DecoratorsSet;
-  oas2?: Oas2DecoratorsSet;
-};
-
-export type TypesExtensionFn = (
-  types: Record<string, NodeType>,
-  oasVersion: OasVersion,
-) => Record<string, NodeType>;
-
-export type TypeExtensionsConfig = Partial<Record<OasMajorVersion, TypesExtensionFn>>;
-export type CustomRulesConfig = {
-  oas3?: Oas3RuleSet;
-  oas2?: Oas2RuleSet;
-};
-
-export type Plugin = {
-  id: string;
-  configs?: Record<string, LintRawConfig>;
-  rules?: CustomRulesConfig;
-  preprocessors?: PreprocessorsConfig;
-  decorators?: DecoratorsConfig;
-  typeExtension?: TypeExtensionsConfig;
-};
-
-export type ResolveHeader =
-  | {
-      name: string;
-      envVariable?: undefined;
-      value: string;
-      matches: string;
-    }
-  | {
-      name: string;
-      value?: undefined;
-      envVariable: string;
-      matches: string;
-    };
-
-export type RawResolveConfig = {
-  http?: Partial<HttpResolveConfig>;
-};
-
-export type HttpResolveConfig = {
-  headers: ResolveHeader[];
-  customFetch?: Function;
-};
-
-export type ResolveConfig = {
-  http: HttpResolveConfig;
-};
-
 export const DEFAULT_REGION = 'us';
-export type Region = 'us' | 'eu';
-export type AccessTokens = { [region in Region]?: string };
 const REDOCLY_DOMAIN = process.env.REDOCLY_DOMAIN;
 export const DOMAINS: { [region in Region]: string } = {
   us: 'redocly.com',
@@ -135,35 +46,6 @@ if (REDOCLY_DOMAIN === 'redoc.online') {
   DOMAINS[REDOCLY_DOMAIN as Region] = REDOCLY_DOMAIN;
 }
 export const AVAILABLE_REGIONS = Object.keys(DOMAINS) as Region[];
-
-export type DeprecatedRawConfig = {
-  apiDefinitions?: Record<string, string>;
-  lint?: LintRawConfig;
-  resolve?: RawResolveConfig;
-  region?: Region;
-  referenceDocs?: Record<string, any>;
-};
-
-export type Api = {
-  root: string;
-  lint?: Omit<LintRawConfig, 'plugins'>;
-  'features.openapi'?: Record<string, any>;
-  'features.mockServer'?: Record<string, any>;
-};
-
-export type RawConfig = {
-  apis?: Record<string, Api>;
-  lint?: LintRawConfig;
-  resolve?: RawResolveConfig;
-  region?: Region;
-  'features.openapi'?: Record<string, any>;
-  'features.mockServer'?: Record<string, any>;
-  organization?: string;
-};
-
-export type ResolvedLintRawConfig = Omit<LintRawConfig, 'extends'> & {
-  extends?: (string | LintRawConfig)[];
-};
 
 export class LintConfig {
   plugins: Plugin[];
@@ -440,285 +322,4 @@ export class Config {
     this.region = rawConfig.region;
     this.organization = rawConfig.organization;
   }
-}
-
-function resolvePresets(presets: string[], plugins: Plugin[]) {
-  return presets.map((presetName) => {
-    const { pluginId, configName } = parsePresetName(presetName);
-    const plugin = plugins.find((p) => p.id === pluginId);
-    if (!plugin) {
-      throw new Error(`Invalid config ${red(presetName)}: plugin ${pluginId} is not included.`);
-    }
-
-    const preset = plugin.configs?.[configName]!;
-    if (!preset) {
-      throw new Error(
-        pluginId
-          ? `Invalid config ${red(
-              presetName,
-            )}: plugin ${pluginId} doesn't export config with name ${configName}.`
-          : `Invalid config ${red(presetName)}: there is no such built-in config.`,
-      );
-    }
-    return preset;
-  });
-}
-
-function parsePresetName(presetName: string): { pluginId: string; configName: string } {
-  if (presetName.indexOf('/') > -1) {
-    const [pluginId, configName] = presetName.split('/');
-    return { pluginId, configName };
-  } else {
-    return { pluginId: '', configName: presetName };
-  }
-}
-
-function resolvePlugins(plugins: (string | Plugin)[] | null, configPath: string = ''): Plugin[] {
-  if (!plugins) return [];
-
-  // @ts-ignore
-  const requireFunc = typeof __webpack_require__ === 'function' ? __non_webpack_require__ : require;
-
-  const seenPluginIds = new Map<string, string>();
-
-  return plugins
-    .map((p) => {
-      // TODO: resolve npm packages similar to eslint
-      const pluginModule =
-        typeof p === 'string'
-          ? (requireFunc(path.resolve(path.dirname(configPath), p)) as Plugin)
-          : p;
-
-      const id = pluginModule.id;
-      if (typeof id !== 'string') {
-        throw new Error(red(`Plugin must define \`id\` property in ${blue(p.toString())}.`));
-      }
-
-      if (seenPluginIds.has(id)) {
-        const pluginPath = seenPluginIds.get(id)!;
-        throw new Error(
-          red(
-            `Plugin "id" must be unique. Plugin ${blue(p.toString())} uses id "${blue(
-              id,
-            )}" already seen in ${blue(pluginPath)}`,
-          ),
-        );
-      }
-
-      seenPluginIds.set(id, p.toString());
-
-      const plugin: Plugin = {
-        id,
-        ...(pluginModule.configs ? { configs: pluginModule.configs } : {}),
-        ...(pluginModule.typeExtension ? { typeExtension: pluginModule.typeExtension } : {}),
-      };
-
-      if (pluginModule.rules) {
-        if (!pluginModule.rules.oas3 && !pluginModule.rules.oas2) {
-          throw new Error(`Plugin rules must have \`oas3\` or \`oas2\` rules "${p}.`);
-        }
-        plugin.rules = {};
-        if (pluginModule.rules.oas3) {
-          plugin.rules.oas3 = prefixRules(pluginModule.rules.oas3, id);
-        }
-        if (pluginModule.rules.oas2) {
-          plugin.rules.oas2 = prefixRules(pluginModule.rules.oas2, id);
-        }
-      }
-      if (pluginModule.preprocessors) {
-        if (!pluginModule.preprocessors.oas3 && !pluginModule.preprocessors.oas2) {
-          throw new Error(
-            `Plugin \`preprocessors\` must have \`oas3\` or \`oas2\` preprocessors "${p}.`,
-          );
-        }
-        plugin.preprocessors = {};
-        if (pluginModule.preprocessors.oas3) {
-          plugin.preprocessors.oas3 = prefixRules(pluginModule.preprocessors.oas3, id);
-        }
-        if (pluginModule.preprocessors.oas2) {
-          plugin.preprocessors.oas2 = prefixRules(pluginModule.preprocessors.oas2, id);
-        }
-      }
-
-      if (pluginModule.decorators) {
-        if (!pluginModule.decorators.oas3 && !pluginModule.decorators.oas2) {
-          throw new Error(`Plugin \`decorators\` must have \`oas3\` or \`oas2\` decorators "${p}.`);
-        }
-        plugin.decorators = {};
-        if (pluginModule.decorators.oas3) {
-          plugin.decorators.oas3 = prefixRules(pluginModule.decorators.oas3, id);
-        }
-        if (pluginModule.decorators.oas2) {
-          plugin.decorators.oas2 = prefixRules(pluginModule.decorators.oas2, id);
-        }
-      }
-
-      return plugin;
-    })
-    .filter(notUndefined);
-}
-
-function prefixRules<T extends Record<string, any>>(rules: T, prefix: string) {
-  if (!prefix) return rules;
-
-  const res: any = {};
-  for (const name of Object.keys(rules)) {
-    res[`${prefix}/${name}`] = rules[name];
-  }
-
-  return res;
-}
-
-type RulesFields =
-  | 'rules'
-  | 'oas2Rules'
-  | 'oas3_0Rules'
-  | 'oas3_1Rules'
-  | 'preprocessors'
-  | 'oas2Preprocessors'
-  | 'oas3_0Preprocessors'
-  | 'oas3_1Preprocessors'
-  | 'decorators'
-  | 'oas2Decorators'
-  | 'oas3_0Decorators'
-  | 'oas3_1Decorators';
-
-function mergeExtends(rulesConfList: LintRawConfig[]) {
-  const result: Omit<LintRawConfig, RulesFields> & Required<Pick<LintRawConfig, RulesFields>> = {
-    rules: {},
-    oas2Rules: {},
-    oas3_0Rules: {},
-    oas3_1Rules: {},
-
-    preprocessors: {},
-    oas2Preprocessors: {},
-    oas3_0Preprocessors: {},
-    oas3_1Preprocessors: {},
-
-    decorators: {},
-    oas2Decorators: {},
-    oas3_0Decorators: {},
-    oas3_1Decorators: {},
-  };
-
-  for (let rulesConf of rulesConfList) {
-    if (rulesConf.extends) {
-      throw new Error(
-        `\`extends\` is not supported in shared configs yet: ${JSON.stringify(
-          rulesConf,
-          null,
-          2,
-        )}.`,
-      );
-    }
-
-    Object.assign(result.rules, rulesConf.rules);
-    Object.assign(result.oas2Rules, rulesConf.oas2Rules);
-    assignExisting(result.oas2Rules, rulesConf.rules || {});
-    Object.assign(result.oas3_0Rules, rulesConf.oas3_0Rules);
-    assignExisting(result.oas3_0Rules, rulesConf.rules || {});
-    Object.assign(result.oas3_1Rules, rulesConf.oas3_1Rules);
-    assignExisting(result.oas3_1Rules, rulesConf.rules || {});
-
-    Object.assign(result.preprocessors, rulesConf.preprocessors);
-    Object.assign(result.oas2Preprocessors, rulesConf.oas2Preprocessors);
-    assignExisting(result.oas2Preprocessors, rulesConf.preprocessors || {});
-    Object.assign(result.oas3_0Preprocessors, rulesConf.oas3_0Preprocessors);
-    assignExisting(result.oas3_0Preprocessors, rulesConf.preprocessors || {});
-    Object.assign(result.oas3_1Preprocessors, rulesConf.oas3_1Preprocessors);
-    assignExisting(result.oas3_1Preprocessors, rulesConf.preprocessors || {});
-
-    Object.assign(result.decorators, rulesConf.decorators);
-    Object.assign(result.oas2Decorators, rulesConf.oas2Decorators);
-    assignExisting(result.oas2Decorators, rulesConf.decorators || {});
-    Object.assign(result.oas3_0Decorators, rulesConf.oas3_0Decorators);
-    assignExisting(result.oas3_0Decorators, rulesConf.decorators || {});
-    Object.assign(result.oas3_1Decorators, rulesConf.oas3_1Decorators);
-    assignExisting(result.oas3_1Decorators, rulesConf.decorators || {});
-  }
-
-  return result;
-}
-
-function assignExisting<T>(target: Record<string, T>, obj: Record<string, T>) {
-  for (let k of Object.keys(obj)) {
-    if (target.hasOwnProperty(k)) {
-      target[k] = obj[k];
-    }
-  }
-}
-
-export function getMergedConfig(config: Config, entrypointAlias?: string): Config {
-  return entrypointAlias
-    ? new Config(
-        {
-          ...config.rawConfig,
-          lint: getMergedLintConfig(config, entrypointAlias),
-          'features.openapi': {
-            ...config['features.openapi'],
-            ...config.apis[entrypointAlias]?.['features.openapi'],
-          },
-          'features.mockServer': {
-            ...config['features.mockServer'],
-            ...config.apis[entrypointAlias]?.['features.mockServer'],
-          },
-          // TODO: merge everything else here
-        },
-        config.configFile,
-      )
-    : config;
-}
-
-export function getMergedLintConfig(config: Config, entrypointAlias?: string) {
-  const apiLint = entrypointAlias ? config.apis[entrypointAlias]?.lint : {};
-  const mergedLint = {
-    ...config.rawConfig.lint,
-    ...apiLint,
-    rules: { ...config.rawConfig.lint?.rules, ...apiLint?.rules },
-    preprocessors: { ...config.rawConfig.lint?.preprocessors, ...apiLint?.preprocessors },
-    decorators: { ...config.rawConfig.lint?.decorators, ...apiLint?.decorators },
-  };
-  return mergedLint;
-}
-
-function transformApiDefinitionsToApis(
-  apiDefinitions: Record<string, string> = {},
-): Record<string, Api> {
-  let apis: Record<string, Api> = {};
-  for (const [apiName, apiPath] of Object.entries(apiDefinitions)) {
-    apis[apiName] = { root: apiPath };
-  }
-  return apis;
-}
-
-export function transformConfig(rawConfig: DeprecatedRawConfig | RawConfig): RawConfig {
-  if ((rawConfig as RawConfig).apis && (rawConfig as DeprecatedRawConfig).apiDefinitions) {
-    throw new Error("Do not use 'apiDefinitions' field. Use 'apis' instead.\n");
-  }
-  if (
-    (rawConfig as RawConfig)['features.openapi'] &&
-    (rawConfig as DeprecatedRawConfig).referenceDocs
-  ) {
-    throw new Error("Do not use 'referenceDocs' field. Use 'features.openapi' instead.\n");
-  }
-  const { apiDefinitions, referenceDocs, ...rest } = rawConfig as DeprecatedRawConfig & RawConfig;
-  if (apiDefinitions) {
-    process.stderr.write(
-      `The ${yellow('apiDefinitions')} field is deprecated. Use ${green(
-        'apis',
-      )} instead. Read more about this change: https://redocly.com/docs/api-registry/guides/migration-guide-config-file/#changed-properties\n`,
-    );
-  }
-  if (referenceDocs) {
-    process.stderr.write(
-      `The ${yellow('referenceDocs')} field is deprecated. Use ${green(
-        'features.openapi',
-      )} instead. Read more about this change: https://redocly.com/docs/api-registry/guides/migration-guide-config-file/#changed-properties\n`,
-    );
-  }
-  return {
-    'features.openapi': referenceDocs,
-    apis: transformApiDefinitionsToApis(apiDefinitions),
-    ...rest,
-  };
 }
