+++ packages/core/src/config/minimal.ts
@@ -1,4 +1,4 @@
-import { LintRawConfig } from './config';
+import type { LintRawConfig } from './types';
 
 export default {
   rules: {
