+++ packages/core/src/config/recommended.ts
@@ -1,4 +1,4 @@
-import { LintRawConfig } from './config';
+import type { LintRawConfig } from './types';
 
 export default {
   rules: {
